revert: remove all changes to @contentful/f36-pill
This PR should only introduce @contentful/f36-pill-next. Deprecation
of the existing Pill and migration guidance will be handled in a
separate PR.
