+++ packages/components/pill/README.mdx
@@ -1,14 +1,12 @@
 ---
-title: 'Pill (Legacy)'
+title: 'Pill'
 type: 'component'
-status: 'deprecated'
+status: 'stable'
 slug: /components/pill/
 github: 'https://github.com/contentful/forma-36/tree/main/packages/components/pill'
 storybook: 'https://f36-storybook.contentful.com/?path=/story/components-pill--basic'
-typescript: ./src/LegacyPill.tsx
+typescript: ./src/Pill.tsx
 ---
-
-> **Deprecated:** This component is deprecated and will be removed in the next major version. Use [`PillNext`](/components/pill-next/) from `@contentful/f36-pill-next` instead.
 
 `Pill` component allows to represent a set of one or multiple objects which can be entered or changed by the user. For example, tags. `Pill` can be integrated with a drag-and-drop solution.
